pyproject.toml
[project]
name = "whisper-api"
description = ""
version = "0.0.1"

dependencies=[
  "alembic ==1.9.0",
  "fastapi ==0.88.0",
  "psycopg2 ==2.9.5",
  "python-dotenv ==0.21.0",
  "sqlalchemy[mypy] == 1.4.45",
  "uvicorn[standard] ==0.20.0"
]

[project.optional-dependencies]
dev = [
  # code formatting
  "black",
  "isort",

  # linting
  "flake8",
  "mypy",

  # tests
  "httpx",
  "sqlalchemy-stubs",
  "sqlalchemy-utils",
  "pytest"
]

[tool.isort]
profile = "black"
